+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changelog for Charlatan
 =======================
 
+0.2.4 (unreleased)
+------------------
+
+- Empty models are allowed so that dict ands lists can be used as fixtures.
+- Fixtures can now inherits from other fixtures.
+
 0.2.3 (2013-06-28)
 ------------------